cloud_functions/collect_options_data/requirements.txt
google-cloud-storage>=2.0.0
google-cloud-secret-manager>=2.0.0
pandas>=1.3.0
polygon-api-client>=1.0.0
pytz>=2021.1
py_vollib>=1.0.0
py_lets_be_rational>=1.0.0
dataclasses>=0.8